适配-1端口

+++ sdk/models.go
@@ -9,11 +9,11 @@ import "encoding/json"
 // Tunnel 隧道
 type Tunnel struct {
 	ID               string `json:"tunnelId"`         // 隧道 ID，8 位小写 Base32
-	Name             string `json:"name"`            // 隧道名称
-	Description      string `json:"description"`     // 隧道描述
-	ExpirationHours  int    `json:"expirationHours"` // 有效期（小时）
+	Name             string `json:"name"`             // 隧道名称
+	Description      string `json:"description"`      // 隧道描述
+	ExpirationHours  int    `json:"expirationHours"`  // 有效期（小时）
 	TunnelExpiration uint32 `json:"tunnelExpiration"` // 隧道过期时间（Unix 秒）
-	PortCount        int    `json:"portCount"`       // 端口数量
+	PortCount        int    `json:"portCount"`        // 端口数量
 }
 
 // TunnelDetail 隧道详情（含状态）
@@ -36,8 +36,8 @@ type TunnelStatus struct {
 // Port 端口配置
 type Port struct {
 	TunnelID       string `json:"tunnelId"`
-	Port           uint16 `json:"port"`
-	Protocol       string `json:"protocol"`        // http, https, auto
+	Port           int32  `json:"port"`
+	Protocol       string `json:"protocol"`       // http, https, auto
 	AllowAnonymous bool   `json:"allowAnonymous"` // 是否允许匿名访问
 }
